Refactored LeafletMap.js to use props, renamed the leaflet map story file to the correct name, uncommented out the calls to the leaflet map story files
