+++ src/Maps/LeafletMap.js
@@ -3,17 +3,17 @@ import { Map, TileLayer } from 'react-leaflet';
 import map from './LeafletMap.styles.css';
 // import . from '../../assets/leaflet.css';
 
-const data = {
-  position: [45.51, -122.68],
-  zoom: 13,
-  // url: "http://{s}.tile.osm.org/{z}/{x}/{y}.png",
-  // attribution: '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors',
-};
+// const data = {
+//   position: [45.51, -122.68],
+//   zoom: 13,
+//   // url: "http://{s}.tile.osm.org/{z}/{x}/{y}.png",
+//   // attribution: '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors',
+// };
 
-const LeafletMap = () => {
+const LeafletMap = (props) => {
   require('../../assets/leaflet.css');
   return (
-    <Map id="map" center={data.position} zoom={data.zoom}>
+    <Map id="map" center={props.center} zoom={props.zoom}>
       <TileLayer
         url="http://{s}.tile.osm.org/{z}/{x}/{y}.png"
         attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
@@ -36,8 +36,8 @@ LeafletMap.displayName = 'LeafletMap';
 LeafletMap.propTypes = {
   center: React.PropTypes.array,
   zoom: React.PropTypes.number,
-  url: React.PropTypes.string,
-  attribution: React.PropTypes.string,
+  // url: React.PropTypes.string,
+  // attribution: React.PropTypes.string,
 };
 
 export default LeafletMap;
